refactor: implement test cases for the appReducer
